fix in data pipeline db

+++ diabetic_retinopathy/input_pipeline/datasets.py
@@ -232,9 +232,10 @@ def prepare_dataset(dataset, augmentation, batch_size, caching, repeat):
 
     # keras model for data augmentation
     data_augmentation = tf.keras.Sequential([
-        RandomChance(tf.keras.layers.RandomFlip("horizontal_and_vertical"),
-                     0.75),
-        RandomChance(tf.keras.layers.RandomRotation(0.5), 0.5)
+        RandomChance(tf.keras.layers.RandomFlip("horizontal_and_vertical",
+                                                dtype='uint8'), 0.75),
+        RandomChance(tf.keras.layers.RandomRotation(0.5, dtype='uint8'),
+                     0.5)
     ])
 
     if caching:
